Formalize the licensing situation
Add a proper license notice to every file in the project, as is
recommended for GPLed code
(https://www.gnu.org/licenses/gpl-faq.html#NoticeInSourceFile).
For files where this is impractical (autogenerated) or impossible
(format does not allow it), add licensing information to the file
`.reuse/dep5`, which is a machine-readable format
(https://dep-team.pages.debian.net/deps/dep5/).

Proper licensing of the entire project can be verified by running `reuse
lint` (https://github.com/fsfe/reuse-tool). This helper tool also
facilitates adding properly formatted headers with license notices to
new files as well as add authors to existing files. Go to
https://reuse.software for more information.

The exception to the above are the following files:
* prefpane/ImageAndTextCell.h
* prefpane/ImageAndTextCell.m
* prefpane/MAAttachedWindow.h
* prefpane/MAAttachedWindow.m

The former two are from ancient example code by Apple. A source for the
original still needs to be located in order to extract the correct
license information.

The latter two are from Matt Gemmell, whom we will contact to sort out
the intended license.

Signed-off-by: Crazor <128295+Crazor@users.noreply.github.com>